+++ R/get_timeseries.R
@@ -3,39 +3,27 @@
 # Concatenate time series environmental data
 # 
 # @param object a numeric vector of geographic coordinates (lonlat) or
-# a matrix with environmental data from other sources.
-# When lonlat is used, the function makes a call to
-# \code{nasapower::get_power} to fetch and combine enviromental data from NASA POWER
-# (\url{https://power.larc.nasa.gov/})
+# a matrix with environmental data from other sources. When lonlat is used, the function makes a call to
+# \code{nasapower} to fetch and combine enviromental data from NASA POWER (\url{https://power.larc.nasa.gov/})
 # @param day.one a vector of class 'Date' for the starting date to capture the environmental data
-# @param span an integer or a vector with integers for the duration of the timespan to be captured
-# @param days.before optional, an integer for the number of days before
-# start.date to be included in the timespan
+# @param span an integer or a vector with integers for the duration of the time series to be captured
+# @param days.before optional, an integer for the number of days before start.date to be included in the timespan
 # @param ... additional arguments passed to \code{\link[nasapower]{get_power}}
-# @return a data frame of environmental data for the chosen period
-#' @import nasapower
-#' @importFrom raster rasterFromXYZ extract stack
-.get_timespan <- function(object, day.one = NULL,
-                          span = NULL, days.before = NULL, ...)
+# @return a data frame of time series environmental data for the chosen period
+# @import nasapower
+.get_timeseries <- function(object, day.one = NULL,
+                            span = 150, days.before = 0, ...)
 {
   
   if (is.null(day.one)) {
-    stop("argument 'day.one' is missing with no default \n")
-  }
-  
-  if (is.null(span)) {
-    stop("argument 'span' is missing with no default \n ")
-  }
-  
-  if (is.null(days.before)) {
-    days.before <- 0
+    stop("Argument 'day.one' is missing with no default \n")
   }
   
   if (is.matrix(object)) {
     object <- as.data.frame(object)
   }
   
-  if (any(c("tbl_df", "tbl") %in% class(day.one))) {
+  if (.is_tibble(day.one)) {
     day.one <- day.one[[1]]
   }
   
@@ -60,8 +48,11 @@
   # look if nasapower is required 
   nasa_power <- dim(object)[[2]] == 2
   
-  # then get data from NASA POWER using nasapower::get_power
+  # then get data from NASA POWER using nasapower
   if (nasa_power) {
+   
+    # the first and last date to fetch
+    dates <- c(min(b), maxend)
     
     # define geographic boundaries for lonlat
     lims <- with(object, c(floor(min(object[,1])), 
@@ -69,10 +60,17 @@
                            ceiling(max(object[,1])), 
                            ceiling(max(object[,2]))))
     
-    # the first and last date to fetch
-    dates <- c(min(b), maxend)
-    # the projection
-    myproj <- "+proj=eqc +lat_ts=0 +lat_0=0 +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
+    # check distance between lims
+    # nasapower accepts a max of 4.5
+    area <- c((lims[3] - lims[1]), 
+              (lims[4] - lims[2]))
+    
+    largearea <- any(area > 4.5)
+    
+    if (largearea) {
+      stop("nasapower supports a maximum area of 4.5 x 4.5 degrees, yours is ", 
+           paste(area, collapse = " x "))
+    }
     
     # get NASA POWER
     info <- nasapower::get_power(community = "AG",
@@ -81,51 +79,57 @@
                                  temporal_average = "DAILY", 
                                  ...)
     
+    class(info) <- class(info)[-1]
+    
+    # rename target fetched product
+    names(info)[ncol(info)] <- "value"
+    
     # split by YYYYMMDD to create a list of data frames
     info <- split(info, info$YYYYMMDD)
     
     # keep only coordinates and the variable fetched
     info <- lapply(info, function(x) {
-      x[(!names(x) %in% c("YEAR", "MM", "DD", "DOY", "YYYYMMDD"))]
+      x[(!names(x) %in% c("YEAR", "MM", "DD", "DOY"))]
     })
     
-    # create a list of raster bricks from each YYYYMMDD data frame
-    r <- lapply(info, function(x) {
-      raster::rasterFromXYZ(x, res = c(0.5, 0.5), crs = myproj)
+    
+    # put this information in its right lonlat as provided in the input
+    xy2 <- info[[1]][,c("LON","LAT")]
+    xy2 <- as.data.frame(xy2)
+    
+    # split lonlat into a list by its rows
+    xy1 <- split(object, 1:nrow(object))
+    
+    # get the index for lonlat in info
+    nn <- lapply(xy1, function(n) {
+      n <- as.vector(t(n))
+      .nearest(xy1 = n, xy2 = xy2)
     })
     
-    # stack the raster
-    r <- raster::stack(unlist(r))
+    # unlist to get the vector
+    nn <- unlist(nn)
     
-    #extract the information using lonlat points provided in 'object'
-    object <- raster::extract(r, object)
+    # force the vector to be in the right order, from 1 to n 
+    nn <- nn[ sort(as.numeric(names(nn))) ]
     
-    object <- as.data.frame(object)
+    # retrieve the data from info using nn
+    data <- lapply(info, function(n) {
+      n <- n[nn, "value"]
+      n
+    })
+    
+    # combine vectors in this list 
+    data <- do.call("cbind", data)
+    
+    object <- as.data.frame(data)
     
     names(object) <- as.character(names(info))
-    
+     
   }
   
   # # check if provided start.date exists within the object
   days <- names(object)
   
-  out <- !as.character(b) %in% days & !as.character(e) %in% days
-
-  if (any(out)) {
-    suppress <- dimnames(object)[[1]][out]
-    warning("Ignoring rows c(",
-            toString(suppress),
-            "): subscript out of bounds \n")
-    rnames <- as.integer(dimnames(object)[[1]][!out])
-  } else{
-    rnames <- as.integer(dimnames(object)[[1]])
-  }
-
-  # refresh objects
-  object <- object[!out, ]
-  b <- b[!out]
-  span <- span[!out]
-  maxspan <- max(span)
   n <- dim(object)[1]
   rownames(object) <- 1:n
   
@@ -162,6 +166,7 @@
   return(Y)
   
 }
+